+++ myproject/myproject-system/src/main/java/com/example/service/SysRoleMenuService.java
@@ -16,10 +16,10 @@ public interface SysRoleMenuService extends IService<SysRoleMenu> {
     List<SysRoleMenu> listRoleMenuByRoleId(Long roleId);
 
 
-    void addRoleMenu(Long roleId, List<Long> menuIds);
+    void addRoleMenu(Long roleId, List<Integer> menuIds);
 
     void deleteRoleMenu(Long roleId);
 
-    void updateRoleMenu(Long roleId, List<Long> menuIds);
+    void updateRoleMenu(Long roleId, List<Integer> menuIds);
 
 }
